Handling HP and death
Added hp points to our character, added dmg to the enemy and defined that when the enemy is in collision with a character it will do dmg to it.
We're displaying health now and lowering it each frame if the character is in contact with the enemy. If hp 0 then stop drawing and game over pops.